backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
pgvector==0.2.4
aiosqlite==0.19.0
python-multipart==0.0.6

# AI Service Integrations (Phase 2)
openai==1.51.0
anthropic==0.39.0
elevenlabs==1.9.0
deepgram-sdk==3.7.2
httpx==0.27.0
certifi==2024.8.30

# CV Analysis (Phase 3)
PyPDF2==3.0.1
python-docx==1.1.0
reportlab==4.0.7

